Wire up handout linking code.
This uses Font Awesome to create a link to the handouts. The links work
heuristically; if there's a "data-handout" attribute it will use that
directly (useful for people with multiple handouts or non-ascii names),
and if not it will use the first author's name in lowercase to look up
the proper handout.